moved functions from main.rs to modules compile and compile_reverse
